+++ src/main/java/net/silentchaos512/gear/gear/trait/TraitSerializers.java
@@ -47,6 +47,7 @@ public final class TraitSerializers {
         register(TargetEffectTrait.SERIALIZER);
         register(BonusDropsTrait.SERIALIZER);
         register(CancelEffectsTrait.SERIALIZER);
+        register(SelfRepairTrait.SERIALIZER);
     }
 
     private TraitSerializers() {}